gerrit_add_repo: Properly fetch username & use it in project.config
Since we don't use groups for single entities (i. e. users) anymore,
we can't use the normal group setup to retrieve the necessary information.

One option would have been to extract it from the MySQL db directly.
And be fucked if/when we switch the db backend. Plus: God knows where
the data will be stored in the future.

Instead, we're now using the gsql interface to Gerrit which *should*
always give us the correct data.

This is the first commit of the Gerrit-2.10/Github fixes.